unify the unit tests.
